refactor(capture): complete capture_internal v0->v1 cutover
Remove the flag-gated dispatch layer and legacy v0 capture client.
The v1 implementation (batch-native, typed options, per-event results)
now owns the capture_internal/capture_batch_internal API names directly.

All 9 capture-v1-* feature flags were already at 100% in production.
Flag deletion in the PostHog app is a separate ops step.
