+++ posthog/api/llm_prompt.py
@@ -15,7 +15,7 @@ from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.serializers import BaseSerializer
 
-from posthog.api.capture_dispatch import capture_internal_routed
+from posthog.api.capture import capture_internal
 from posthog.api.llm_prompt_serializers import (
     LLMPromptDuplicateSerializer,
     LLMPromptFetchQuerySerializer,
@@ -192,7 +192,7 @@ class LLMPromptViewSet(
     def _track_prompt_fetch(self, prompt: dict[str, Any]) -> None:
         if not settings.TEST:
             try:
-                capture_internal_routed(
+                capture_internal(
                     token=self.team.api_token,
                     event_name=PROMPT_FETCHED_EVENT,
                     event_source=PROMPT_FETCHED_EVENT_SOURCE,
