+++ products/conversations/backend/api/tests/test_events.py
@@ -40,7 +40,7 @@ class TestConversationEvents(BaseTest):
             anonymous_traits={"name": "Test Customer", "email": "test@example.com"},
         )
 
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     def test_capture_ticket_created_uses_team_token(self, mock_capture):
         capture_ticket_created(self.ticket)
 
@@ -54,7 +54,7 @@ class TestConversationEvents(BaseTest):
         assert call_kwargs["properties"]["customer_name"] == "Test Customer"
         assert call_kwargs["properties"]["customer_email"] == "test@example.com"
 
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     def test_capture_ticket_status_changed_uses_team_token(self, mock_capture):
         capture_ticket_status_changed(self.ticket, "new", "pending", actor=self.user, actor_type="user")
 
@@ -70,7 +70,7 @@ class TestConversationEvents(BaseTest):
         assert call_kwargs["properties"]["actor_id"] == self.user.id
         assert call_kwargs["properties"]["actor_email"] == self.user.email
 
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     def test_capture_ticket_priority_changed_uses_team_token(self, mock_capture):
         capture_ticket_priority_changed(self.ticket, None, "high", actor=self.user, actor_type="user")
 
@@ -86,7 +86,7 @@ class TestConversationEvents(BaseTest):
         assert call_kwargs["properties"]["actor_id"] == self.user.id
         assert call_kwargs["properties"]["actor_email"] == self.user.email
 
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     def test_capture_ticket_assigned_uses_team_token(self, mock_capture):
         capture_ticket_assigned(self.ticket, "user", "123", actor=self.user, actor_type="user")
 
@@ -102,7 +102,7 @@ class TestConversationEvents(BaseTest):
         assert call_kwargs["properties"]["actor_id"] == self.user.id
         assert call_kwargs["properties"]["actor_email"] == self.user.email
 
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     def test_capture_message_sent_uses_team_token(self, mock_capture):
         capture_message_sent(self.ticket, "msg-123", "Hello customer", author=self.user)
 
@@ -129,7 +129,7 @@ class TestConversationEvents(BaseTest):
             ("capture_message_sent", capture_message_sent, ["msg-id", "content"]),
         ]
     )
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     def test_customer_email_falls_back_to_email_from(self, _name, capture_fn, extra_args, mock_capture):
         self.ticket.anonymous_traits = {}
         self.ticket.email_from = "customer@example.com"
@@ -142,7 +142,7 @@ class TestConversationEvents(BaseTest):
         call_kwargs = mock_capture.call_args.kwargs
         assert call_kwargs["properties"]["customer_email"] == "customer@example.com"
 
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     def test_capture_message_received_uses_team_token(self, mock_capture):
         capture_message_received(self.ticket, "msg-456", "Hello support")
 
@@ -183,7 +183,7 @@ class TestConversationEvents(BaseTest):
             ),
         ]
     )
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     def test_all_events_include_base_properties(self, _name, capture_fn, expected_event, extra_args, mock_capture):
         capture_fn(self.ticket, *extra_args)
 
@@ -197,7 +197,7 @@ class TestConversationEvents(BaseTest):
         assert props["status"] == self.ticket.status
         assert props["priority"] == self.ticket.priority
 
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     def test_message_content_truncated_to_1000_chars(self, mock_capture):
         long_content = "x" * 1500
         capture_message_sent(self.ticket, "msg-id", long_content, author=self.user)
@@ -205,7 +205,7 @@ class TestConversationEvents(BaseTest):
         call_kwargs = mock_capture.call_args.kwargs
         assert len(call_kwargs["properties"]["message_content"]) == 1000
 
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     def test_event_uses_ticket_team_token_not_other_team(self, mock_capture):
         """Verify events route to the ticket's team, not any other team."""
         from posthog.models import Organization, Team
@@ -221,7 +221,7 @@ class TestConversationEvents(BaseTest):
         assert call_kwargs["token"] == self.team.api_token
         assert call_kwargs["token"] != other_team.api_token
 
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     def test_two_teams_events_routed_to_respective_projects(self, mock_capture):
         """Events from Team 1 ticket use Team 1 token, Team 2 ticket uses Team 2 token."""
         from posthog.models import Organization, Team
@@ -263,7 +263,7 @@ class TestConversationEvents(BaseTest):
             ("empty_distinct_id", False, False, False),
         ]
     )
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     @patch("products.conversations.backend.events.get_persons_by_distinct_ids")
     def test_capture_ticket_created_person_processing(
         self, _name, has_person, is_identified, expect_groups, mock_get_persons, mock_capture
@@ -296,7 +296,7 @@ class TestConversationEvents(BaseTest):
         assert call_kwargs["process_person_profile"] is expect_groups
         assert "$groups" not in call_kwargs["properties"]
 
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     @patch("products.conversations.backend.events.get_persons_by_distinct_ids")
     def test_capture_ticket_created_groups_from_person_org(self, mock_get_persons, mock_capture):
         from posthog.models.person.person import Person
@@ -316,7 +316,7 @@ class TestConversationEvents(BaseTest):
         assert groups["project"] == str(self.team.uuid)
         assert "instance" in groups
 
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     @patch("products.conversations.backend.events.get_persons_by_distinct_ids")
     def test_capture_message_received_groups_from_person_org(self, mock_get_persons, mock_capture):
         from posthog.models.person.person import Person
@@ -343,7 +343,7 @@ class TestConversationEvents(BaseTest):
             ("empty_distinct_id", False, False),
         ]
     )
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     @patch("products.conversations.backend.events.get_persons_by_distinct_ids")
     def test_capture_message_received_person_processing(
         self, _name, has_person, is_identified, mock_get_persons, mock_capture
@@ -376,7 +376,7 @@ class TestConversationEvents(BaseTest):
         assert call_kwargs["process_person_profile"] is False
         assert "$groups" not in call_kwargs["properties"]
 
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     @patch("products.conversations.backend.events.get_persons_by_distinct_ids", side_effect=Exception("db timeout"))
     def test_capture_message_received_still_fires_on_person_lookup_failure(self, mock_get_persons, mock_capture):
         capture_message_received(self.ticket, "msg-456", "Hello support")
@@ -392,7 +392,7 @@ class TestConversationEvents(BaseTest):
             ("user_has_no_membership", True, False),
         ]
     )
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     @patch("products.conversations.backend.events.get_persons_by_distinct_ids")
     def test_capture_ticket_created_no_groups_when(
         self, _name, create_user, create_membership, mock_get_persons, mock_capture
@@ -410,7 +410,7 @@ class TestConversationEvents(BaseTest):
         assert call_kwargs["process_person_profile"] is False
         assert "$groups" not in call_kwargs["properties"]
 
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     @patch("products.conversations.backend.events.get_persons_by_distinct_ids", side_effect=Exception("db timeout"))
     def test_capture_ticket_created_still_fires_on_person_lookup_failure(self, mock_get_persons, mock_capture):
         capture_ticket_created(self.ticket)
@@ -429,7 +429,7 @@ class TestConversationEvents(BaseTest):
             ("email_from_only", "email", None, "biz@acme.com"),
         ]
     )
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     @patch("products.conversations.backend.person_lookup._get_persons_by_email")
     @patch("products.conversations.backend.events.get_persons_by_distinct_ids")
     def test_capture_ticket_created_email_fallback_groups(
@@ -478,7 +478,7 @@ class TestConversationEvents(BaseTest):
             ("person_without_membership", True),
         ]
     )
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     @patch("products.conversations.backend.person_lookup._get_persons_by_email")
     @patch("products.conversations.backend.events.get_persons_by_distinct_ids")
     def test_capture_ticket_created_email_fallback_no_groups(
@@ -512,7 +512,7 @@ class TestConversationEvents(BaseTest):
         assert call_kwargs["process_person_profile"] is False
         assert "$groups" not in call_kwargs["properties"]
 
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     @patch("products.conversations.backend.person_lookup._get_persons_by_email")
     @patch("products.conversations.backend.events.get_persons_by_distinct_ids")
     def test_capture_ticket_created_identified_person_without_membership_skips_email_fallback(
@@ -538,7 +538,7 @@ class TestConversationEvents(BaseTest):
             ("github_no_email", "github"),
         ]
     )
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     @patch("products.conversations.backend.person_lookup._get_persons_by_email")
     @patch("products.conversations.backend.events.get_persons_by_distinct_ids")
     def test_capture_ticket_created_non_verified_channels_skip_email_fallback(
@@ -578,7 +578,7 @@ class TestConversationEvents(BaseTest):
             ("without_customer", ""),
         ]
     )
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     @patch("posthog.hogql.query.execute_hogql_query")
     @patch("products.conversations.backend.events.get_group_types_for_project")
     @patch("products.conversations.backend.events.get_persons_by_distinct_ids")
@@ -616,7 +616,7 @@ class TestConversationEvents(BaseTest):
             ("empty_org_key", [["", ""]]),
         ]
     )
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     @patch("posthog.hogql.query.execute_hogql_query")
     @patch("products.conversations.backend.events.get_group_types_for_project")
     @patch("products.conversations.backend.events.get_persons_by_distinct_ids")
@@ -635,7 +635,7 @@ class TestConversationEvents(BaseTest):
         assert call_kwargs["process_person_profile"] is False
         assert "$groups" not in call_kwargs["properties"]
 
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     @patch("posthog.hogql.query.execute_hogql_query")
     @patch("products.conversations.backend.events.get_persons_by_distinct_ids")
     def test_capture_ticket_created_membership_hit_skips_analytics_fallback(
@@ -655,7 +655,7 @@ class TestConversationEvents(BaseTest):
         call_kwargs = mock_capture.call_args.kwargs
         assert call_kwargs["properties"]["$groups"]["organization"] == str(person_org.id)
 
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     @patch("posthog.hogql.query.execute_hogql_query")
     @patch("products.conversations.backend.events.get_group_types_for_project")
     @patch("products.conversations.backend.person_lookup._get_persons_by_email")
@@ -695,7 +695,7 @@ class TestConversationEvents(BaseTest):
             ("negative", [], False),
         ]
     )
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     @patch("posthog.hogql.query.execute_hogql_query")
     @patch("products.conversations.backend.events.get_group_types_for_project")
     @patch("products.conversations.backend.events.get_persons_by_distinct_ids")
@@ -719,7 +719,7 @@ class TestConversationEvents(BaseTest):
             else:
                 assert "$groups" not in call.kwargs["properties"]
 
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     def test_capture_ticket_status_changed_system_actor_uses_customer_distinct_id(self, mock_capture):
         """System actions (e.g. snooze wake) use the customer's distinct_id, not an actor's."""
         capture_ticket_status_changed(self.ticket, "on_hold", "open", actor_type="system")
@@ -731,7 +731,7 @@ class TestConversationEvents(BaseTest):
         assert call_kwargs["properties"]["actor_id"] is None
         assert call_kwargs["properties"]["actor_email"] is None
 
-    @patch("products.conversations.backend.events.capture_internal_routed")
+    @patch("products.conversations.backend.events.capture_internal")
     def test_capture_ticket_status_changed_external_actor_uses_customer_distinct_id(self, mock_capture):
         """External actions (e.g. GitHub sync) use the customer's distinct_id."""
         capture_ticket_status_changed(self.ticket, "open", "resolved", actor_type="external")
