+++ products/conversations/backend/events.py
@@ -10,7 +10,7 @@ from typing import Literal
 
 import structlog
 
-from posthog.api.capture_dispatch import capture_internal_routed
+from posthog.api.capture import capture_internal
 from posthog.clickhouse.query_tagging import Feature, Product, tags_context
 from posthog.event_usage import groups as build_groups
 from posthog.models.group_type_mapping import get_group_types_for_project
@@ -249,7 +249,7 @@ def capture_ticket_created(ticket: Ticket) -> None:
     except Exception:
         logger.exception("ticket_created_person_lookup_failed", team_id=team_id, ticket_id=str(ticket.id))
 
-    capture_internal_routed(
+    capture_internal(
         token=team.api_token,
         event_name="$conversation_ticket_created",
         event_source=EVENT_SOURCE,
@@ -272,7 +272,7 @@ def capture_ticket_status_changed(
     properties["new_status"] = new_status
     properties.update(_get_actor_properties(actor, actor_type))
 
-    capture_internal_routed(
+    capture_internal(
         token=ticket.team.api_token,
         event_name="$conversation_ticket_status_changed",
         event_source=EVENT_SOURCE,
@@ -294,7 +294,7 @@ def capture_ticket_priority_changed(
     properties["new_priority"] = new_priority
     properties.update(_get_actor_properties(actor, actor_type))
 
-    capture_internal_routed(
+    capture_internal(
         token=ticket.team.api_token,
         event_name="$conversation_ticket_priority_changed",
         event_source=EVENT_SOURCE,
@@ -316,7 +316,7 @@ def capture_ticket_assigned(
     properties["assignee_id"] = assignee_id
     properties.update(_get_actor_properties(actor, actor_type))
 
-    capture_internal_routed(
+    capture_internal(
         token=ticket.team.api_token,
         event_name="$conversation_ticket_assigned",
         event_source=EVENT_SOURCE,
@@ -340,7 +340,7 @@ def capture_message_sent(
     properties.update(_get_actor_properties(author, "user"))
     properties.update(_get_customer_properties(ticket, include_distinct_id=True))
 
-    capture_internal_routed(
+    capture_internal(
         token=ticket.team.api_token,
         event_name="$conversation_message_sent",
         event_source=EVENT_SOURCE,
@@ -367,7 +367,7 @@ def capture_message_received(ticket: Ticket, message_id: str, message_content: s
     except Exception:
         logger.exception("message_received_person_lookup_failed", team_id=team.id, ticket_id=str(ticket.id))
 
-    capture_internal_routed(
+    capture_internal(
         token=team.api_token,
         event_name="$conversation_message_received",
         event_source=EVENT_SOURCE,
